requirements.txt
beautifulsoup4==4.5.0
boto3==1.9.48
chardet==3.0.4
Django==2.2.14
django-bootstrap3-datepicker==0.4.1
django-boto==0.3.12
django-compressor==2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.7
django-storages==1.7.1

django-taggit==0.23.0
djangorestframework==3.8.2
django-celery==3.2.2
django-redis==4.8.0
django-cors-headers==2.3.0
gunicorn==19.9.0
Markdown==2.6.9
pytz==2017.3
psycopg2==2.7.3.2
PyPDF2==1.26.0
python-dateutil==2.6.1
pylint-django==2.1.0
pusher==1.7.2
pypinyin==0.30.0
reportlab==3.4.0
redis==2.10.6
six==1.10.0
urllib3==1.22
requests==2.18.4
wechat-sdk==0.6.4
django-sass-processor==0.7.2
libsass==0.16.1
easy-thumbnails==2.5
mutagen==1.42.0
audioread==2.1.6
xmltodict==0.9.2
selenium==3.141.0
tenacity==5.0.3
xlwt==1.3.0
xlrd==1.2.0
xlrd==1.2.0
xlutils==2.0.0
python-gitlab==1.10.0
django-multiselectfield==0.1.8
aliyun-python-sdk-core==2.13.10
django-filter==2.2.0
django-crum==0.7.6
oss2==2.12.1
json-tools==0.4.1
python-docx==0.8.10
channels==2.2
channels-redis==2.4.2
daphne==2.3
docxcompose==1.3.1
pdfminer==20191125
pdfplumber==0.5.27